refactoring: move shared code from vtreader to feature_helper
