+++ feature_helper.py
@@ -4,6 +4,7 @@ from file_helper import FileHelper
 from log_helper import info, warn, critical, debug
 import processing
 import uuid
+import numbers
 
 
 class FeatureMerger:
@@ -81,3 +82,89 @@ class FeatureMerger:
         processing.runalg("qgis:dissolve", layer, False, "dissolveGroup", target_file)
         dissolved_layer = QgsVectorLayer(target_file, "Dissolved", "ogr")
         return dissolved_layer
+
+
+tile_extent = 4096  # this applies always for Mapbox tiles (see: https://github.com/tilezen/mapbox-vector-tile)
+
+
+class _GeoTypes:
+    def __init__(self):
+        pass
+
+    POINT = "Point"
+    LINE_STRING = "LineString"
+    POLYGON = "Polygon"
+
+GeoTypes = _GeoTypes()
+
+geo_types = {
+    1: GeoTypes.POINT,
+    2: GeoTypes.LINE_STRING,
+    3: GeoTypes.POLYGON}
+
+
+def is_multi(geo_type, coordinates):
+    """
+    * Returns true, if the specified coordinates belong to a Multi geometry (e.g. MultiPolygon or MultiLineString)
+    :param geo_type:
+    :param coordinates:
+    :return:
+    """
+
+    if geo_type == GeoTypes.POINT:
+        is_single = len(coordinates) == 2 and all(isinstance(c, int) for c in coordinates)
+        return not is_single
+    elif geo_type == GeoTypes.LINE_STRING:
+        is_array_of_tuples = all(len(c) == 2 and all(isinstance(ci, int) for ci in c) for c in coordinates)
+        is_single = is_array_of_tuples
+        return not is_single
+    elif geo_type == GeoTypes.POLYGON:
+        return get_array_depth(coordinates, 0) >= 2
+    return False
+
+
+def get_array_depth(arr, depth):
+    """
+    * Returns the depth of an array.
+      >> Example: arr=[1,2,3], depth=0, then the resulting depth will be 0
+      >> Example: arr=[[1,2], [3,4]], depth=0, then the resulting depth will be 1
+    :param arr:
+    :param depth:
+    :return:
+    """
+
+    if all(isinstance(c, numbers.Real) for c in arr[0]):
+        return depth
+    else:
+        depth += 1
+        return get_array_depth(arr[0], depth)
+
+
+def map_coordinates_recursive(coordinates, mapper_func, all_out_of_bounds_func=None):
+    """
+    Recursively traverses the array of coordinates (depth first) and applies the specified function
+    """
+    any_tuples_inside_bounds = False
+    tuple_count_on_current_array_depth = 0
+    tmp = []
+    is_coordinate_tuple = len(coordinates) == 2 and all(isinstance(c, int) for c in coordinates)
+    if is_coordinate_tuple:
+        newval = mapper_func(coordinates)
+        tmp.append(newval)
+    else:
+        for coord in coordinates:
+            is_coordinate_tuple = len(coord) == 2 and all(isinstance(c, int) for c in coord)
+            if is_coordinate_tuple:
+                tuple_count_on_current_array_depth += 1
+                if not any_tuples_inside_bounds and 1 <= coord[0] <= tile_extent and 1 <= coord[1] <= tile_extent:
+                    any_tuples_inside_bounds = True
+
+                newval = mapper_func(coord)
+                tmp.append(newval)
+            else:
+                tmp.append(map_coordinates_recursive(coord, mapper_func, all_out_of_bounds_func))
+
+    all_out_of_bounds = tuple_count_on_current_array_depth > 0 and not any_tuples_inside_bounds
+    if all_out_of_bounds_func:
+        all_out_of_bounds_func(all_out_of_bounds)
+    return tmp
